requirements.txt
flask
flask-pymongo
flask-jwt-extended
flask-cors
flask-bcrypt
pymongo
python-dotenv
gunicorn
google-auth
pyotp
google-auth-oauthlib
google-auth-httplib2
